Complete react functionality for the header
I figured out how to set up react to create the structure for the header. The tricky part was figuring how to get all of the files to interact. The importing and exporting was a bit confusing at first. I haven't completed the styling, but that shouldn't take much time.