Complete project for first React Module
I got so caught up in figuring things out that I completely forgot to keep my commits up to date. This contains all I've done for the project. Style the footer, but leave out functionality cause I was unclear of what was being asked. Looked up moment.js and figured out how to include the current date as part of the title.